adding some changes in my server above all the update subject, and also creating new request for my new tables (blogs and publications)

+++ FinalProject/CaminoAlbaProject/user_manager_android_app_project-main/user-management-spring-boot/user-management/src/main/java/com/user_manager_v1/repository/ProfileRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProfileRepository extends JpaRepository<Profile,Long> {
+public interface ProfileRepository extends JpaRepository<Profile,Integer> {
 
     @Query(value = "SELECT person_id FROM person WHERE person_id = :email ", nativeQuery = true)
     List<String> checkUserEmail(@Param("email") String email);
